+++ desc/magnetic_fields/_current_potential.py
@@ -16,14 +16,14 @@ from ._core import _MagneticField, biot_savart_general
 class CurrentPotentialField(_MagneticField, FourierRZToroidalSurface):
     """Magnetic field due to a surface current potential on a toroidal surface.
 
-        surface current K is assumed given by
-         K = n x ∇ Φ
-        where:
-               n is the winding surface unit normal.
-               Phi is the current potential function,
-                which is a function of theta and zeta.
-        This function then uses biot-savart to find the
-        B field from this current density K on the surface.
+    Surface current K is assumed given by K = n x ∇ Φ
+    where:
+
+        - n is the winding surface unit normal.
+        - Phi is the current potential function, which is a function of theta and zeta.
+
+    This function then uses biot-savart to find the B field from this current
+    density K on the surface.
 
     Parameters
     ----------
@@ -159,7 +159,7 @@ class CurrentPotentialField(_MagneticField, FourierRZToroidalSurface):
     def save(self, file_name, file_format=None, file_mode="w"):
         """Save the object.
 
-        **Not supported for this object!
+        **Not supported for this object!**
 
         Parameters
         ----------
@@ -275,19 +275,21 @@ class FourierCurrentPotentialField(
 ):
     """Magnetic field due to a surface current potential on a toroidal surface.
 
-        surface current K is assumed given by
+    Surface current K is assumed given by
 
-        K = n x ∇ Φ
-        Φ(θ,ζ) = Φₛᵥ(θ,ζ) + Gζ/2π + Iθ/2π
+    K = n x ∇ Φ
 
-        where:
-              n is the winding surface unit normal.
-              Phi is the current potential function,
-                which is a function of theta and zeta,
-                and is given as a secular linear term in theta/zeta
-                and a double Fourier series in theta/zeta.
-        This function then uses biot-savart to find the
-        B field from this current density K on the surface.
+    Φ(θ,ζ) = Φₛᵥ(θ,ζ) + Gζ/2π + Iθ/2π
+
+    where:
+
+        - n is the winding surface unit normal.
+        - Phi is the current potential function, which is a function of theta and zeta,
+          and is given as a secular linear term in theta/zeta and a double Fourier
+          series in theta/zeta.
+
+    This function then uses biot-savart to find the B field from this current
+    density K on the surface.
 
     Parameters
     ----------