+++ tests/conftest.py
@@ -169,7 +169,7 @@ def HELIOTRON_vac(tmpdir_factory):
     print("cwd=", cwd)
 
     args = ["-o", str(desc_h5_path), input_filename, "-vv"]
-    with pytest.warns(UserWarning, match="Vacuum objective does not use any profiles"):
+    with pytest.warns(UserWarning, match="Vacuum objective assumes 0 pressure"):
         main(args)
 
     HELIOTRON_vacuum_out = {