+++ tests/test_objective_funs.py
@@ -19,7 +19,11 @@ from desc.equilibrium import Equilibrium
 from desc.examples import get
 from desc.geometry import FourierRZToroidalSurface
 from desc.grid import ConcentricGrid, LinearGrid, QuadratureGrid
-from desc.magnetic_fields import FourierCurrentPotentialField, SplineMagneticField
+from desc.magnetic_fields import (
+    FourierCurrentPotentialField,
+    OmnigenousField,
+    SplineMagneticField,
+)
 from desc.objectives import (
     AspectRatio,
     BootstrapRedlConsistency,
@@ -37,6 +41,7 @@ from desc.objectives import (
     MercierStability,
     ObjectiveFromUser,
     ObjectiveFunction,
+    Omnigenity,
     PlasmaVesselDistance,
     Pressure,
     PrincipalCurvature,
@@ -446,7 +451,7 @@ class TestObjectiveFunction:
         eq.solve()
         obj = BoundaryError(eq, coilset, field_grid=coil_grid)
         obj.build()
-        f = obj.compute_scaled_error(*obj.xs(eq))
+        f = obj.compute_scaled_error(*obj.xs())
         n = len(f) // 3
         # first n should be B*n errors
         np.testing.assert_allclose(f[:n], 0, atol=1e-4)
@@ -465,7 +470,7 @@ class TestObjectiveFunction:
         eq.solve()
         obj = BoundaryError(eq, coilset, field_grid=coil_grid)
         obj.build()
-        f = obj.compute_scaled_error(*obj.xs(eq))
+        f = obj.compute_scaled_error(*obj.xs())
         n = len(f) // 2
         # first n should be B*n errors
         np.testing.assert_allclose(f[:n], 0, atol=1e-4)
@@ -482,7 +487,7 @@ class TestObjectiveFunction:
         eq.solve()
         obj = VacuumBoundaryError(eq, coilset, field_grid=coil_grid)
         obj.build()
-        f = obj.compute_scaled_error(*obj.xs(eq))
+        f = obj.compute_scaled_error(*obj.xs())
         n = len(f) // 2
         # first n should be B*n errors
         np.testing.assert_allclose(f[:n], 0, atol=1e-4)
@@ -499,7 +504,7 @@ class TestObjectiveFunction:
         eq.solve()
         obj = BoundaryErrorNESTOR(eq, coilset, field_grid=coil_grid)
         obj.build()
-        f = obj.compute_scaled_error(*obj.xs(eq))
+        f = obj.compute_scaled_error(*obj.xs())
         np.testing.assert_allclose(f, 0, atol=2e-3)
 
     @pytest.mark.unit
@@ -789,7 +794,7 @@ def test_derivative_modes():
 
 @pytest.mark.unit
 def test_generic_compute():
-    """Test for gh issue #388."""
+    """Test for GH issue #388."""
     eq = Equilibrium()
     obj = ObjectiveFunction(AspectRatio(target=2, weight=1, eq=eq), use_jit=False)
     obj.build()
@@ -1061,7 +1066,7 @@ def test_boundary_error_print(capsys):
     n = len(f) // 2
     f1 = f[:n]
     f2 = f[n:]
-    obj.print_value(*obj.xs(eq))
+    obj.print_value(*obj.xs())
     out = capsys.readouterr()
 
     corr_out = str(
@@ -1136,7 +1141,7 @@ def test_boundary_error_print(capsys):
     n = len(f) // 2
     f1 = f[:n]
     f2 = f[n:]
-    obj.print_value(*obj.xs(eq))
+    obj.print_value(*obj.xs())
     out = capsys.readouterr()
 
     corr_out = str(
@@ -1213,7 +1218,7 @@ def test_boundary_error_print(capsys):
     f1 = f[:n]
     f2 = f[n : 2 * n]
     f3 = f[2 * n :]
-    obj.print_value(*obj.xs(eq))
+    obj.print_value(*obj.xs())
     out = capsys.readouterr()
 
     corr_out = str(
@@ -1756,6 +1761,41 @@ class TestComputeScalarResolution:
             f[i] = obj.compute_scalar(obj.x())
         np.testing.assert_allclose(f, f[-1], rtol=5e-2)
 
+    # TODO: add test for PlasmaVesselDistance
+
+    # Omnigenity
+    # (needs an approximately QP equilibrium and field)
+    surf = FourierRZToroidalSurface.from_qp_model(
+        major_radius=1,
+        aspect_ratio=20,
+        elongation=6,
+        mirror_ratio=0.2,
+        torsion=0.1,
+        NFP=1,
+        sym=True,
+    )
+    eq = Equilibrium(Psi=6e-3, M=4, N=4, surface=surf)
+    eq, _ = eq.solve(objective="force", verbose=3)
+    field = OmnigenousField(
+        L_B=0,
+        M_B=2,
+        L_x=0,
+        M_x=0,
+        N_x=0,
+        NFP=eq.NFP,
+        helicity=(0, eq.NFP),
+        B_lm=np.array([0.8, 1.2]),
+    )
+    f = np.zeros_like(res_array, dtype=float)
+    for i, res in enumerate(res_array + 0.5):  # omnigenity needs higher resolution
+        grid = LinearGrid(M=int(eq.M * res), N=int(eq.N * res), NFP=eq.NFP)
+        obj = ObjectiveFunction(
+            Omnigenity(eq=eq, field=field, eq_grid=grid, field_grid=grid)
+        )
+        obj.build(verbose=0)
+        f[i] = obj.compute_scalar(obj.x(eq, field))
+    np.testing.assert_allclose(f, f[-1], rtol=1e-3)
+
 
 class TestObjectiveNaNGrad:
     """Make sure reverse mode AD works correctly for all objectives."""
@@ -1780,6 +1820,8 @@ class TestObjectiveNaNGrad:
         GenericObjective,
         LinearObjectiveFromUser,
         ObjectiveFromUser,
+        # TODO: add Omnigenity objective (see GH issue #943)
+        Omnigenity,
     ]
     other_objectives = list(set(objectives) - set(specials))