+++ tests/test_plotting.py
@@ -19,6 +19,7 @@ from desc.examples import get
 from desc.geometry import FourierRZToroidalSurface, FourierXYZCurve
 from desc.grid import ConcentricGrid, Grid, LinearGrid, QuadratureGrid
 from desc.io import load
+from desc.magnetic_fields import OmnigenousField
 from desc.plotting import (
     plot_1d,
     plot_2d,
@@ -767,11 +768,7 @@ def test_plot_boozer_surface():
     fig, ax, data = plot_boozer_surface(
         eq, M_booz=eq.M, N_booz=eq.N, return_data=True, rho=0.5, fieldlines=4
     )
-    for string in [
-        "|B|",
-        "theta_Boozer",
-        "zeta_Boozer",
-    ]:
+    for string in ["|B|", "theta_B", "zeta_B"]:
         assert string in data.keys()
     return fig
 
@@ -905,4 +902,23 @@ def test_plot_3d_surface():
     """Test 3d plotting of surface object."""
     surf = FourierRZToroidalSurface()
     fig = plot_3d(surf, "curvature_H_rho")
+    return fig
+
+
+@pytest.mark.unit
+@pytest.mark.mpl_image_compare(remove_text=True, tolerance=tol_2d)
+def test_plot_omnigenous_field():
+    """Test plot omnigenous magnetic field."""
+    field = OmnigenousField(
+        L_B=0,
+        M_B=4,
+        L_x=0,
+        M_x=1,
+        N_x=1,
+        NFP=4,
+        helicity=(1, 4),
+        B_lm=np.array([0.8, 0.9, 1.1, 1.2]),
+        x_lmn=np.array([0, -np.pi / 8, 0, np.pi / 8, 0, np.pi / 4]),
+    )
+    fig, ax = plot_boozer_surface(field, iota=0.6, fieldlines=4)
     return fig